+++ packages/flutter_blue_plus_android/lib/flutter_blue_plus_android.dart
@@ -5,12 +5,13 @@ import 'package:flutter/foundation.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_blue_plus_platform_interface/flutter_blue_plus_platform_interface.dart';
 
-class FlutterBluePlusAndroid extends FlutterBluePlusPlatform {
+final class FlutterBluePlusAndroid extends FlutterBluePlusPlatform {
   @visibleForTesting
   final methodChannel = const MethodChannel('flutter_blue_plus/methods');
 
   var _initialized = false;
   var _logLevel = LogLevel.none;
+  var _logColor = true;
 
   final _onAdapterStateChangedController = StreamController<BmBluetoothAdapterState>.broadcast();
   final _onBondStateChangedController = StreamController<BmBondStateResponse>.broadcast();
@@ -127,7 +128,7 @@ class FlutterBluePlusAndroid extends FlutterBluePlusPlatform {
   ) async {
     return await _invokeMethod<bool>(
       'createBond',
-      request.remoteId.str,
+      request.toMap(),
     ) == true;
   }
 
@@ -292,6 +293,7 @@ class FlutterBluePlusAndroid extends FlutterBluePlusPlatform {
     BmSetLogLevelRequest request,
   ) async {
     _logLevel = request.logLevel;
+    _logColor = request.logColor;
 
     return await _invokeMethod<bool>(
       'setLogLevel',
@@ -395,18 +397,26 @@ class FlutterBluePlusAndroid extends FlutterBluePlusPlatform {
 
     // log args
     if (_logLevel == LogLevel.verbose) {
-      print("[FBP] <$method> args: $arguments");
+      var func = '<$method>';
+      var args = arguments.toString();
+      func = _logColor ? '\x1B[1;30m$func\x1B[0m' : func;
+      args = _logColor ? '\x1B[1;35m$args\x1B[0m' : args;
+      print('[FBP] $func args: $args');
     }
 
     // invoke
-    final result = await methodChannel.invokeMethod<T>(method, arguments);
+    final out = await methodChannel.invokeMethod<T>(method, arguments);
 
     // log result
     if (_logLevel == LogLevel.verbose) {
-      print("[FBP] ($method) result: $result");
+      var func = '($method)';
+      var result = out.toString();
+      func = _logColor ? '\x1B[1;30m$func\x1B[0m' : func;
+      result = _logColor ? '\x1B[1;33m$result\x1B[0m' : result;
+      print('[FBP] $func result: $result');
     }
 
-    return result;
+    return out;
   }
 
   Future<void> _initFlutterBluePlus() async {
@@ -433,12 +443,14 @@ class FlutterBluePlusAndroid extends FlutterBluePlusPlatform {
   ) async {
     // log result
     if (_logLevel == LogLevel.verbose) {
-      if (call.method == 'OnDiscoveredServices') {
-        // this is really slow so we can't pretty print anything that happens a lot
-        print('[FBP] [[ ${call.method} ]] result: ${_prettyPrint(call.arguments)}');
-      } else {
-        print('[FBP] [[ ${call.method} ]] result: ${call.arguments}');
-      }
+      var func = '[[ ${call.method} ]]';
+      var result = switch (call.method) {
+        'OnDiscoveredServices' => _prettyPrint(call.arguments),
+        _ => call.arguments.toString(),
+      };
+      func = _logColor ? '\x1B[1;30m$func\x1B[0m' : func;
+      result = _logColor ? '\x1B[1;33m$result\x1B[0m' : result;
+      print('[FBP] $func result: $result');
     }
 
     // handle method call