+++ packages/flutter_blue_plus_android/CHANGELOG.md
@@ -1,2 +1,11 @@
+## 3.0.0
+* Add option to provide pairing PIN to `createBond` (#1119)
+
+## 2.0.1
+* Add log color
+
+## 2.0.0
+* Replace void return types with bool return types
+
 ## 1.35.0
-* Splits from `flutter_blue_plus` as a federated implementation
+* Split from `flutter_blue_plus` as a federated implementation