+++ packages/flutter_blue_plus_linux/CHANGELOG.md
@@ -1,2 +1,8 @@
+## 3.0.0
+* Update platform interface version to 3.0.0
+
+## 2.0.0
+* Replace void return types with bool return types
+
 ## 1.0.0
 * Initial release