+++ packages/flutter_blue_plus_platform_interface/CHANGELOG.md
@@ -1,2 +1,14 @@
+## 3.0.0
+* Add support for web optional services (#1124)
+* Add option to provide pairing PIN to `createBond` (#1119)
+
+## 2.0.1
+* Add log color
+
+## 2.0.0
+* Replace plugin_platform_interface with base keyword
+* Return sensible defaults instead of throwing
+* Replace void return types with bool return types
+
 ## 1.0.0
 * Initial release