+++ packages/flutter_blue_plus_web/CHANGELOG.md
@@ -1,2 +1,11 @@
+## 3.0.1
+* Fixed getAdapterState() implementation
+
+## 3.0.0
+* Add support for web optional services (#1124)
+
+## 2.0.0
+* Replace void return types with bool return types
+
 ## 1.0.0
 * Initial release